address pr review comment, added cleanup, updated regex in ci.yml
